test: delete-clients
