+++ performance-testing/src/delete-clients_{client_id}.test.ts
@@ -1,0 +1,18 @@
+import { Options } from 'k6/options';
+
+import { ClientGateway } from './gateway';
+import { dummyCreateClientRequest } from "./dummy";
+
+import client from './client';
+
+export let options: Options = {
+  vus: 50,
+  duration: '10s'
+};
+
+const clientGateway = new ClientGateway(client);
+
+export default () => {
+  const client = clientGateway.create(dummyCreateClientRequest());
+  clientGateway.delete(client.id);
+};
